♻️ Created component for the AI
